Implemented Animations & EventSystem & Const variables for Clean Code
Implemented Animations which include picking, walking and idle. Used Event System to make decoupled from visuality and logic. In Testing NavMesh Script. I fired EventSystem on the function which includes calculations of NPC distance. In the Animator class, I subscribe to Event and SetBool to  run the animations according to calculations.